Makefile: Remove hardcoded setuptools version
Since Python 3.10, there was improvement for collections.abc [1] and
basically you should use collections.abc instead of collections [2].

Otherwise, reForis setup.py fails to install, because of outdated and
hardcoded setuptools version.

It fails with this error:
        File "/foo/reforis/venv/lib/python3.10/site-packages/pkg_resources/_vendor/pyparsing.py", line 943, in <module>
          collections.MutableMapping.register(ParseResults)
      AttributeError: module 'collections' has no attribute 'MutableMapping'

Based on the error, I was able to track it down that setuptools needs
to be updated to at least version 40.4.3 [3], which includes a newer version of
pyparsing with Python 3.10 support. [4]

Because of that, there were two ways how to solve it:
a) update setuptools version to more recent one
b) drop the hardcoded version as it seems that translations now works
with new setuptools versions (confirmed by Aleks)

[1] https://docs.python.org/3/whatsnew/3.10.html#collections-abc
[2] https://bugzilla.mozilla.org/show_bug.cgi?id=1718878
[3] https://github.com/pypa/setuptools/commit/2862b90422c63afc8864952878594e4194c517d9
[4] https://github.com/pyparsing/pyparsing/releases/tag/pyparsing_2.2.1