gitlab-ci: Add test for JavaScript files on node 16
We need to make sure that reForis works with node 14 and node 16.
Both these versions are supported by Node.js.

The thing is that OpenWrt 21.02 is using node 14, but in Turris OS 6.0,
which is based on that OpenWrt version, there is a difference, that we
are using node 16, which is the default in OpenWrt 22.03, but OpenWrt
master is using node 18.

We could drop test for node 14, but we need to be sure that reForis
works under two node versions as in the future, we will be using node 16
and node 18 in the Turris OS across multiple OpenWrt versions.